refactor(pat): extract api-keys.tsx into composable files
Split the 851-line settings/api-keys.tsx into a folder of small, focused
pieces. The page file is now a thin composer that picks between the PAT
list and the legacy project key; state, network, and UI details live in
their own components.

- utils.ts                    shared helpers (mask, env-format, role summary)
- CodeBlock.tsx               reusable copy/reveal code block
- ProjectApiKeySection.tsx    legacy sk-lw-* section
- CreatePatDrawer.tsx         owns transient create form state
- TokenCreatedDialog.tsx      one-time token display
- RevokeConfirmDialog.tsx     revoke confirmation
- PersonalAccessTokensSection.tsx  PAT list + orchestrates the dialogs
- index.tsx                   page-level tabs composer
